operations_ai_system/requirements.txt
streamlit>=1.30.0
openai>=1.12.0
pandas>=2.1.0
openpyxl>=3.1.0
plotly>=5.18.0
python-dotenv>=1.0.0
jinja2>=3.1.0
resend>=0.7.0
